git-svn-id: https://svn.apache.org/repos/asf/tomcat/tc7.0.x/tags/TOMCAT_7_0_55@1611608 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/connector/MapperListener.java
@@ -124,6 +124,9 @@ public class MapperListener extends LifecycleMBeanBase
     @Override
     public void stopInternal() throws LifecycleException {
         setState(LifecycleState.STOPPING);
+
+        Engine engine = (Engine) connector.getService().getContainer();
+        removeListeners(engine);
     }
 
 
@@ -156,7 +159,11 @@ public class MapperListener extends LifecycleMBeanBase
                 } else if (child instanceof Context) {
                     registerContext((Context) child);
                 } else if (child instanceof Wrapper) {
-                    registerWrapper((Wrapper) child);
+                    // Only if the Context has started. If it has not, then it
+                    // will have its own "after_start" life-cycle event later.
+                    if (child.getParent().getState().isAvailable()) {
+                        registerWrapper((Wrapper) child);
+                    }
                 }
             }
         } else if (Container.REMOVE_CHILD_EVENT.equals(event.getType())) {
@@ -391,11 +398,6 @@ public class MapperListener extends LifecycleMBeanBase
      */
     private void unregisterContext(Context context) {
 
-        // Don't un-map a context that is paused
-        if (context.getPaused()){
-            return;
-        }
-
         String contextPath = context.getPath();
         if ("/".equals(contextPath)) {
             contextPath = "";
@@ -407,8 +409,23 @@ public class MapperListener extends LifecycleMBeanBase
                     contextPath, connector));
         }
 
-        mapper.removeContextVersion(hostName, contextPath,
-                context.getWebappVersion());
+        if (context.getPaused()) {
+            if (log.isDebugEnabled()) {
+                log.debug(sm.getString("mapperListener.pauseContext",
+                        contextPath, connector));
+            }
+
+            mapper.pauseContextVersion(context, hostName, contextPath,
+                    context.getWebappVersion());
+        } else {
+            if (log.isDebugEnabled()) {
+                log.debug(sm.getString("mapperListener.unregisterContext",
+                        contextPath, connector));
+            }
+
+            mapper.removeContextVersion(hostName, contextPath,
+                    context.getWebappVersion());
+        }
     }
 
 
@@ -482,12 +499,7 @@ public class MapperListener extends LifecycleMBeanBase
             if (obj instanceof Wrapper) {
                 unregisterWrapper((Wrapper) obj);
             } else if (obj instanceof Context) {
-                Context c = (Context) obj;
-                // Only unregister if not paused. If paused, need to keep
-                // registration in place to prevent 404's during reload
-                if (!c.getPaused()) {
-                    unregisterContext(c);
-                }
+                unregisterContext((Context) obj);
             } else if (obj instanceof Host) {
                 unregisterHost((Host) obj);
             }
